+++ Runtime/Core/Http/HttpErrorParser.cs
@@ -178,6 +178,7 @@ namespace AccelByte.Core
             switch (error.Code)
             {
                 case ErrorCode.SessionVersionMismatch:
+                case ErrorCode.AchievementTooManyRequests:
                     return true;
             }